feat: Fazla mesai ve nöbet saati için yeni kurallar ve geliştirmeler
Bu değişiklik, kullanıcı tarafından istenen üç yeni özelliği ve bir hata düzeltmesini içermektedir:

1.  **Fazla Mesai Saatleri Tam Sayı Oldu:** Fazla mesai saatleri artık ondalık değerler yerine yalnızca tam sayılar olarak girilebilir. Bu, hem veritabanı şemasında (`INTEGER` veri türü) hem de kullanıcı arayüzünde (`step="1"`) zorunlu kılındı.
2.  **Nöbet Saati Uyarısı:** Bir çalışanın ay içindeki toplam nöbet saati (hafta sonu ve tatil günleri), toplam çalışılan gün sayısının yarısını aştığında kullanıcıyı uyaran bir sistem eklendi.
3.  **Çalışan Bilgisi Gösterimi:** "Çalışma Saatleri" sekmesinde, çalışanların adlarının yanında artık çalışan kimlik numaraları da gösterilerek daha kolay tanınmaları sağlandı.
4.  **Hata Düzeltmesi:** Uygulamanın başlangıçta `openpyxl` bağımlılığının eksik olması nedeniyle çökmesine neden olan bir sorun giderildi. Bağımlılıklar `requirements.txt` dosyasına eklendi ve yüklendi.

+++ database.py
@@ -28,8 +28,8 @@ def init_db():
             id INTEGER PRIMARY KEY AUTOINCREMENT,
             employee_id INTEGER NOT NULL,
             date TEXT NOT NULL,
-            day_hours REAL DEFAULT 0,
-            evening_hours REAL DEFAULT 0,
+            day_hours INTEGER DEFAULT 0,
+            evening_hours INTEGER DEFAULT 0,
             sunday_reason TEXT,
             FOREIGN KEY (employee_id) REFERENCES employees (id) ON DELETE CASCADE
         )
